#1067 Double -> BigDecimal
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13113 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/opentaps-common/src/common/org/opentaps/common/domain/product/ProductRepository.java
@@ -90,7 +90,7 @@ public class ProductRepository extends Repository implements ProductRepositoryIn
             if (ServiceUtil.isError(results)) {
                 throw new RepositoryException(ServiceUtil.getErrorMessage(results));
             }
-            return BigDecimal.valueOf((Double) results.get("productCost"));
+            return (BigDecimal) results.get("productCost");
         } catch (GenericServiceException e) {
             throw new RepositoryException(e);
         }
